mysite/requirements.txt
django 
djangorestframework
environs
mysqlclient